// packages/core/src/libs/has.d.ts
export type HasCondition<CategoryType> =
	| CategoryType
	| {
			and: HasCondition<CategoryType> | HasCondition<CategoryType>[];
	  }
	| {
			or: HasCondition<CategoryType> | HasCondition<CategoryType>[];
	  }
	| {
			not: HasCondition<CategoryType>;
	  };